Created an abstract class and the queue worker. Simpler
Added file update capabitilies. Processor generates a File entity, the worker adds it to the Node. @giancarlobi need to document and explain how we avoid double processing here. We mass more data around, but it saves us from overprocessing so that is good?
